add dynamic tiles and renderables

+++ oni-core/inc/math/functions.h
@@ -1,12 +1,12 @@
 #pragma once
 
 #define _USE_MATH_DEFINES
-#include <math.h>
+#include <cmath>
 
 namespace oni {
 	namespace math {
 		inline float toRadians(float degrees) {
-			return degrees * (M_PI / 180.0f);
+			return static_cast<float>(degrees * (M_PI / 180.0f));
 		}
 	}
 }
